peview: Fix 32bit build warning

+++ tools/peview/peheaderprp.c
@@ -754,7 +754,7 @@ VOID PvSetPeImageOverlayHeaderProperties(
 
             imageDosStubEntropy = PvCalculateEntropyBuffer(
                 imageOverlayData,
-                imageOverlayDataLength,
+                (ULONG)imageOverlayDataLength,
                 &imageDosStubMean
                 );
